Rename TUSResumableUpload2
