+++ Pod/Classes/TUSSession.h
@@ -7,7 +7,7 @@
 #import <Foundation/Foundation.h>
 #import "TUSUploadStore.h"
 #import "TUSData.h"
-#import "TUSResumableUpload2.h"
+#import "TUSResumableUpload.h"
 
 /**
  Session that manages, creates, and reloads TUS uploads using a single NSURLSession and data store
@@ -27,14 +27,14 @@
 /**
  Create an upload, but do not start it
  */
-- (TUSResumableUpload2 *) createUploadFromFile:(NSURL *)fileURL
+- (TUSResumableUpload *) createUploadFromFile:(NSURL *)fileURL
                                        headers:(NSDictionary <NSString *, NSString *> * __nullable)headers
                                       metadata:(NSDictionary <NSString *, NSString *> * __nullable)metadata;
 
 /**
  Restore an upload, but do not start it.  Uploads must be restored by ID because file URLs can change between launch.
  */
-- (TUSResumableUpload2 *) restoreUpload:(NSString *)uploadId;
+- (TUSResumableUpload *) restoreUpload:(NSString *)uploadId;
 
 //TODO: Allow custom TUSData to be passed in with an upload.
 
@@ -44,7 +44,7 @@
  
  This is not done automatically so that an application can choose which to load into memory.
  */
--(NSArray <TUSResumableUpload2 *> *)restoreAllUploads;
+-(NSArray <TUSResumableUpload *> *)restoreAllUploads;
 
 
 // Cancel all pending uploads
@@ -56,5 +56,5 @@
  
  This is not done automatically on restore so that an application can choose which to resume.
  */
--(NSArray <TUSResumableUpload2 *> *)resumeAll;
+-(NSArray <TUSResumableUpload *> *)resumeAll;
 @end
